+++ packages/flow-core/src/components/f-icon/f-icon.ts
@@ -44,7 +44,8 @@ export class FIcon extends FRoot {
     | "success"
     | "danger"
     | "warning"
-    | "neutral" = "default";
+    | "neutral"
+    | "inherit" = "default";
 
   /**
    * @attribute Source property defines what will be displayed on the icon. For icon variant It can take the icon name from a library , any inline SVG or any URL for the image. For emoji, it takes emoji as inline text.
@@ -133,6 +134,7 @@ export class FIcon extends FRoot {
   render() {
     // validating properties
     this.validateProperties();
+    // if state prop is inherit, this would inherit color properties of its latest parent f-div.
 
     /**
      * Final html to render
@@ -147,9 +149,7 @@ export class FIcon extends FRoot {
     >
       ${this.loading
         ? html`${unsafeSVG(loader)}`
-        : html`${this.isURLSource
-            ? unsafeHTML(this.source)
-            : unsafeSVG(this.source)}`}
+        : html`${this.isURLSource ? unsafeHTML(this.source) : unsafeSVG(this.source)}`}
     </div>`;
   }
 }