+++ packages/flow-core/CHANGELOG.md
@@ -4,15 +4,26 @@
 
 # Change Log
 
-## [1.0.8] - 2022-10-31
+## [1.0.9] - 2022-11-04
 
+### Features
+
+- Inheritance Rules added in various components, where if state is equal to inherit, then the particular component will inherit color of its parent.
+  Following are the components:
+  - `f-div`
+  - `f-icon`
+  - `f-text`
+  - `f-pictogram`
+  - `f-button`
+  - `f-icon-button`
+
+## [1.0.8] - 2022-10-31
 
 ### Improvements
 
 - `README.md` added in npm package.
 
 ## [1.0.7] - 2022-10-31
-
 
 ### Improvements